Renaming files to remove redundant "HAL_" prefix.
