+++ src/HAL/ESP32/ESP32.h
