+++ src/HAL/Mega2560/Mega2560.h
@@ -38,7 +38,7 @@
   #define HAL_SERIAL_TRANSMIT
 
   // Use low overhead serial
-  #include "HAL_Serial.h"
+  #include "Serial.h"
 #else
   // New symbols for the Serial ports so they can be remapped if necessary -----------------------------
   #define SerialA Serial
