+++ src/HAL/STM32F1/STM32F1.ino
